feat()Adding payment End Point

+++ src/main/java/com/api/water_sytem_management_java/controllers/dtos/CustomerInput.java
@@ -9,8 +9,8 @@ public record CustomerInput(
          String contact,
          String address,
          CustomerStatus status,
-         Integer valve,
-         Integer monthsInDebt){
+      byte valve,
+         byte monthsInDebt){
 
     public Customer toCustomer() {
      return  new Customer(name,
